+++ tests/test_env_default.py
@@ -1,0 +1,35 @@
+from marabunta.config import EnvDefault, BoolEnvDefault
+
+import os
+
+
+def test_env_default_str():
+    os.environ['test_var'] = 'Foo'
+    test = EnvDefault('test_var', option_strings='', dest='test_var')
+    assert(test.default == 'Foo')
+
+
+def test_bool_env_default_str():
+    os.environ['test_var'] = 'True'
+    test = BoolEnvDefault('test_var', option_strings='', dest='test_var')
+    assert(test.default)
+
+    os.environ['test_var'] = 'true'
+    test = BoolEnvDefault('test_var', option_strings='', dest='test_var')
+    assert(test.default)
+
+    os.environ['test_var'] = '1'
+    test = BoolEnvDefault('test_var', option_strings='', dest='test_var')
+    assert(test.default)
+
+    os.environ['test_var'] = 'False'
+    test = BoolEnvDefault('test_var', option_strings='', dest='test_var')
+    assert(not test.default)
+
+    os.environ['test_var'] = 'false'
+    test = BoolEnvDefault('test_var', option_strings='', dest='test_var')
+    assert(not test.default)
+
+    os.environ['test_var'] = '3'
+    test = BoolEnvDefault('test_var', option_strings='', dest='test_var')
+    assert(not test.default)